requirements/requirements.txt
aniso8601==8.1.1
appdirs==1.4.4
attrs==20.3.0
Card-Validator==1.0.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
credit-card-checker==0.0.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
httpretty==1.0.5
identify==1.5.13
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nodeenv==1.5.0
packaging==20.9
pluggy==0.13.1
pre-commit==2.10.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-mock==1.8.0
requests-toolbelt==0.9.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.1
Werkzeug==1.0.1
zipp==3.4.0
